Webhook to sync user creation data from Clerk to backend
Updated the readme to document steps on how to setup local testing of the webhook and how to do it in a production environment. But, its important to keep the clerk database in sync with the local db so we can create db relations in our backend and have persisted user data. Less round trips too. Bit of a setup process but got it working with a clerk webhook handler.